feat(heating): replace manual animation with lottie animation
