+++ src/components/store/SocketProviderValue.tsx
@@ -26,7 +26,7 @@ import { OS_UPDATE_STATUS } from '../../hooks/useDeviceOSStatus';
 import { OSStatusResponse } from '@meticulous-home/espresso-api';
 import { useIdleTimer } from '../../hooks/useIdleTimer';
 
-const SERVER_URL: string = window.env?.SERVER_URL ?? 'http://localhost:8080';
+const SERVER_URL: string = window.env?.SERVER_URL || 'http://localhost:8080';
 const socket: Socket | null = io(SERVER_URL);
 
 export const SocketProviderValue = () => {
